+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 ### Added
-- Side pannel for superusers:
+- Side panel for superusers:
   - Work copy selection
   - Swap versions
   - Reassign rooms